pyproject.toml
[tool.poetry]
name = "fast-bp"
version = "0.1.0"
description = "Projeto: API para gerenciamento de Usuarios e Atividades"
authors = ["Urich Oliveira"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
dynaconf = "^3.1.2"
fastapi = "^0.61.1"
python-jose = "^3.2.0"
passlib = "^1.7.4"
SQLAlchemy = "^1.3.20"
arrow = "^0.17.0"
bcrypt = "^3.2.0"
uvicorn = "^0.12.2"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.15.0"
pylint = "^2.6.0"
Hypercorn = "^0.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
